[Feat] Extension 코드 추가 #30
