+++ Carrot-iOS/Carrot-iOS/Network/Foundation/BaseService.swift
@@ -100,14 +100,3 @@ extension BaseService {
         }
     }
 }
-
-
-extension MoyaProvider {
-    func request(_ target: Target) async -> Result<Response, MoyaError> {
-        await withCheckedContinuation { continuation in
-            self.request(target) { result in
-                continuation.resume(returning: result)
-            }
-        }
-    }
-}
